Exported resources now are deleted in the RFS only, if they were deleted in the publishing process. Resources in deleted folders are not listed to be exported any longer.

+++ src/org/opencms/staticexport/A_CmsStaticExportHandler.java
@@ -1,7 +1,7 @@
 /*
  * File   : $Source: /alkacon/cvs/opencms/src/org/opencms/staticexport/A_CmsStaticExportHandler.java,v $
- * Date   : $Date: 2006/11/16 13:01:55 $
- * Version: $Revision: 1.3.4.11 $
+ * Date   : $Date: 2006/11/17 11:10:40 $
+ * Version: $Revision: 1.3.4.12 $
  *
  * This library is part of OpenCms -
  * the Open Source Content Mananagement System
@@ -65,7 +65,7 @@ import org.apache.commons.logging.Log;
  * 
  * @author Michael Emmerich
  * 
- * @version $Revision: 1.3.4.11 $ 
+ * @version $Revision: 1.3.4.12 $ 
  * 
  * @since 6.1.7 
  * 
@@ -234,7 +234,10 @@ public abstract class A_CmsStaticExportHandler implements I_CmsStaticExportHandl
                         vfsName)
                         + rfsName.substring(OpenCms.getStaticExportManager().getRfsPrefix(vfsName).length()));
 
-                    purgeFile(rfsExportFileName, vfsName);
+                    if (res.getState().isDeleted()) {
+                        // physically remove files only, if they are marked as deleted
+                        purgeFile(rfsExportFileName, vfsName);
+                    }
                     scrubedFiles.add(rfsName);
 
                     List fileList = getRelatedFilesToPurge(rfsExportFileName, vfsName);
